refactor(dsg): pass custom-actions ast function to the DSG
