+++ libs/util/code-gen-utils/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@amplication/code-gen-utils",
-  "version": "0.0.9",
+  "version": "0.0.7",
   "dependencies": {
     "@babel/parser": "7.23.0",
     "@babel/types": "7.23.0",
