+++ libs/util/code-gen-utils/src/lib/ast/index.ts
@@ -4,4 +4,3 @@ export * from "./remove-typescript-ignore-comments";
 export * from "./remove-typescript-interface-declares";
 export * from "./remove-typescript-variable-declares";
 export * from "./remove-imports-typescript-ignore-comments";
-export * from "./remove-custom-actions-functions";
