--- libs/util/code-gen-utils/src/lib/ast/remove-custom-actions-functions/index.ts
@@ -1,1 +1,0 @@
-export * from "./main";
