+++ packages/data-service-generator/src/server/resource/test/create-controller-spec.ts
@@ -1,4 +1,4 @@
-import { namedTypes, builders } from "ast-types";
+import { namedTypes, builders, ASTNode } from "ast-types";
 import * as kinds from "ast-types/gen/kinds";
 import {
   print,
@@ -6,9 +6,6 @@ import {
   removeTSClassDeclares,
   removeTSInterfaceDeclares,
   removeTSIgnoreComments,
-  removeObjectPropertyByName,
-  removeCallExpressionStatementByName,
-  removeObjectMethodByName,
 } from "@amplication/code-gen-utils";
 import { camelCase } from "camel-case";
 import replaceExt from "replace-ext";
@@ -32,6 +29,7 @@ import { createServiceId } from "../service/create-service";
 import { createControllerId } from "../controller/create-controller";
 import pluginWrapper from "../../../plugin-wrapper";
 import DsgContext from "../../../dsg-context";
+import { visit } from "recast";
 
 const testTemplatePath = require.resolve("./controller.spec.template.ts");
 const TO_ISO_STRING_ID = builders.identifier("toISOString");
@@ -310,4 +308,70 @@ function createFieldTestValueFromPrisma(
       return null;
     }
   }
+}
+
+function removeCallExpressionStatementByName(
+  ast: ASTNode,
+  action: ModuleAction
+): void {
+  visit(ast, {
+    visitCallExpression(path) {
+      const expression = path.value as namedTypes.CallExpression;
+      const { actionType, name } = action;
+      const argument = expression.arguments.find(
+        (a) => a.type === "StringLiteral"
+      ) as namedTypes.StringLiteral;
+
+      if (!argument) return this.traverse(path);
+
+      if (
+        actionType === EnumModuleActionType.Create &&
+        argument.value.includes("POST")
+      ) {
+        path.prune();
+      }
+
+      if (
+        actionType === EnumModuleActionType.Find &&
+        argument.value === `GET /${name}`
+      ) {
+        path.prune();
+      }
+
+      if (
+        actionType === EnumModuleActionType.Read &&
+        argument.value.includes(":id")
+      ) {
+        path.prune();
+      }
+
+      this.traverse(path);
+    },
+  });
+}
+
+function removeObjectMethodByName(ast: ASTNode, actionName: string): void {
+  visit(ast, {
+    visitObjectMethod(path) {
+      const keyName = path.value.key.name;
+      if (keyName === actionName) {
+        path.prune();
+      }
+
+      this.traverse(path);
+    },
+  });
+}
+
+function removeObjectPropertyByName(ast: ASTNode, actionName: string): void {
+  visit(ast, {
+    visitObjectProperty(path) {
+      const keyName = path.value.key.name;
+      if (keyName === actionName) {
+        path.prune();
+      }
+
+      this.traverse(path);
+    },
+  });
 }
